#95 on_data_change handler was throwing away all but last value so client couldn't see other values on a monitored item with a small monitoring rate with a long publish interval

+++ client/src/session.rs
@@ -431,7 +431,7 @@ impl Session {
                                         client_handle: item.client_handle(),
                                         sampling_interval: item.sampling_interval(),
                                         filter: ExtensionObject::null(),
-                                        queue_size: item.queue_size(),
+                                        queue_size: item.queue_size() as u32,
                                         discard_oldest: true,
                                     },
                                 })
